13 Mar 15
--Got a reply to my google groups post from one of the code authors(Jorg
  Kappes).  He said just to comment out the offending line and not worry about
  it.  Not sure it matters much, since I think it will just be easier to use
  Explicit Function instead of Sparse Function.

--Finished first draft of new OpenGM code.  Throwing a runtime error at this
  point.  Will investigate more.

+++ source/sentence-codetection/inference.h
@@ -1,1 +1,3 @@
-double bp_object_inference(double **f, double **g, int T, int top_k, int dummy_f, int dummy_g, int num_gscores, int *boxes);
+double bp_object_inference(double **f, double **g, int T, int top_k, 
+			   double dummy_f, double dummy_g, 
+			   int num_gscores, int *boxes);
